Move utils files to their own directory
