+++ server/utils/encounterUtils.js
